convert diabetes module to GMF
