+++ lib/entity/entity.rb
@@ -14,11 +14,11 @@ module Synthea
     end
 
     def [](name)
-      @attributes[name]
+      @attributes[name.to_sym]
     end
 
     def []=(name, value)
-      @attributes[name] = value
+      @attributes[name.to_sym] = value
     end
 
     def had_event?(type, time = nil)
